Fix useScopedRedirect hook (#164)
* Fix useScopedRedirect hook

* Refactor user UI

* Move user-me to own directory
