+++ src/containers/internal/layout.tsx
@@ -1,22 +1,18 @@
 import React, { useState } from "react";
-import { Link, Outlet } from "react-router-dom";
+import { Outlet } from "react-router-dom";
 import { Button, Icon, classNames } from "jambonz-ui";
 
+import { UserMe } from "./user-me";
 import { Navi } from "./navi";
 import { Icons } from "src/components";
-import { toastSuccess, useSelectState } from "src/store";
+import { toastSuccess } from "src/store";
 import { useAuth } from "src/router/auth";
 import { useMobileMedia } from "src/utils";
 import { MSG_LOGGED_OUT } from "src/constants";
 
 import "./styles.scss";
-import { ROUTE_INTERNAL_USERS } from "src/router/routes";
-import { useApiData } from "src/api";
-import { CurrentUserData } from "src/api/types";
 
 export const Layout = () => {
-  const user = useSelectState("user");
-  const [userData] = useApiData<CurrentUserData>("Users/me");
   const [active, setActive] = useState(false);
   const { signout } = useAuth();
   const mobile = useMobileMedia();
@@ -48,22 +44,7 @@ export const Layout = () => {
           <Icon subStyle="dark" onClick={handleMenu}>
             <Icons.Menu />
           </Icon>
-          <div className="user">
-            <Icons.User className="user--icon" />
-            <div className="item__info">
-              <div className="user--txt">
-                <Link
-                  to={`${ROUTE_INTERNAL_USERS}/${user?.user_sid}/edit`}
-                  title="Edit user"
-                >
-                  <strong>{userData?.user.name}</strong>
-                </Link>
-              </div>
-              <div>
-                <strong>Scope:</strong> <code>{user?.scope}</code>
-              </div>
-            </div>
-          </div>
+          <UserMe />
           <Button
             small
             mainStyle="hollow"
