+++ src/styles/index.scss
@@ -5,7 +5,6 @@
 @use "./inpbtn";
 @use "./filters";
 @use "./grid";
-@use "./user";
 @use "jambonz-ui/src/styles/index";
 
 @use "./vars";
